Added dynamic workforce scale and hashes for reacting to changes in the files

+++ EnDaBaServices/SafeFileDeleter.cs
@@ -1,13 +1,14 @@
 using System;
 
-namespace ImmichEnDaBa;
+namespace EnDaBaServices;
 
 public static class SafeFileDelete
 {
-    private static string _tempFolder = Path.GetTempPath(); 
+    private static readonly string _tempFolder = Path.GetTempPath(); 
 
-    private static string[] allowedFilesToDelete = [
+    private static readonly string[] allowedFilesToDelete = [
         ".zip",
+        ".hash.txt",
     ];
 
     public static void Delete(string path) 
@@ -16,11 +17,17 @@ public static class SafeFileDelete
             throw new ArgumentException("can only delete files in the temp folder");
         }
 
+        bool hasAllowedExtension = false;
         foreach (var fileExtension in allowedFilesToDelete)
         {
-            if (path.EndsWith(fileExtension) is false) {
-                throw new ArgumentException($"can only delete {fileExtension} files");
+            if (path.EndsWith(fileExtension)) {
+                hasAllowedExtension = true;
+                break;
             }        
+        }
+
+        if (hasAllowedExtension is false) {
+            throw new ArgumentException($"cannot delete files with a {Path.GetExtension(path)} extension");
         }
 
         File.Delete(path);
